separate hello world logic from application
