+++ service/src/main/kotlin/eu/miaplatform/service/core/applications/DocumentationApplication.kt
